requirements.txt
alembic==1.7.4
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
mysqlclient==2.0.3
protobuf==3.19.1
PyMySQL==1.0.2
python-dotenv==0.19.1
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0

redis~=3.5.3